refactor: APIモジュールを複数のサブモジュールに分割
- APIモジュールを機能ごとに複数のサブモジュールに分割
- Promise版とコールバック版の関数を自動生成するヘルパー関数を追加
- 循環参照を解消するために遅延読み込みを導入
- 各サブモジュールに適切なドキュメントコメントを追加
